refactor: move and rename `invalidate_token`
