+++ apps/quickbooks_online/actions.py
@@ -3,8 +3,8 @@ from datetime import datetime, timezone
 
 from django.conf import settings
 from django.db.models import Q
-from apps.exceptions import invalidate_token
-from apps.workspaces.actions import patch_integration_settings
+from fyle_qbo_api.utils import invalidate_qbo_credentials
+from fyle_qbo_api.utils import patch_integration_settings
 from django_q.tasks import Chain
 from fyle_accounting_mappings.models import MappingSetting
 from qbosdk.exceptions import InvalidTokenError, WrongParamsError
@@ -54,7 +54,7 @@ def get_preferences(workspace_id: int):
     except QBOCredential.DoesNotExist:
         return Response(data={'message': 'QBO credentials not found in workspace'}, status=status.HTTP_400_BAD_REQUEST)
     except (WrongParamsError, InvalidTokenError):
-        invalidate_token(workspace_id, qbo_credentials)
+        invalidate_qbo_credentials(workspace_id, qbo_credentials)
         return Response(data={'message': 'Invalid token or Quickbooks Online connection expired'}, status=status.HTTP_400_BAD_REQUEST)
